homework_3/requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.7.1
google-cloud-storage==1.13.0
gunicorn==19.9.0
Flask-PyMongo>=2.0.0
oauth2client==4.1.2
pymongo[tls,srv]==3.7.2
six==1.11.0
google-cloud-vision==0.36.0

